a lottt

+++ vendors/views.py
@@ -12,25 +12,16 @@ from rest_framework.exceptions import ValidationError, PermissionDenied
 # Create your views here.
 @swagger_auto_schema(methods=['POST'] ,
                     request_body=FoodSerializer())
-@api_view(['GET', 'POST'])
+@api_view(['POST'])
 @authentication_classes([BasicAuthentication])
 @permission_classes([IsAuthenticated])
 def food_view(request):
-    user = request.user
-    # get all food in the database
-    if request.method == "GET":
-        food_items = Food.objects.all()
-        serializer = FoodSerializer(food_items, many=True)
-        
-        data = {
-           "message":"successful",
-           "data": serializer.data
-        }
     
-    
-        return Response(data, status=status.HTTP_200_OK)
-    # add food to database and menu
-    elif request.method == 'POST':
+    # add food to database and vendor's menu
+    if request.method == 'POST':
+        user=request.user
+        if user.is_vendor == False:
+            raise PermissionDenied(detail={"message":f"Permission Denied. Only vendors can perform this action"})
         serializer = FoodSerializer(data=request.data)
         if serializer.is_valid(): 
             if "user" in serializer.validated_data.keys():
@@ -48,7 +39,8 @@ def food_view(request):
             Menu.objects.create(food=food,user=user)
             data = {
                 'message' : 'success',
-                'data'  : new_serializer.data
+                'data'  : new_serializer.data,
+                'menu id': Menu.objects.get(food=food,user=user).id
             }
             return Response(data, status=status.HTTP_202_ACCEPTED)
         else:
@@ -71,7 +63,7 @@ def food_view(request):
 @permission_classes([IsAuthenticated])
 def food_detail(request, food_id):
 
-    
+    # get, edit and delete individual food
     try:
         food = Food.objects.get(id=food_id)
     except Food.DoesNotExist:
@@ -81,8 +73,13 @@ def food_detail(request, food_id):
             'error'  : f"Food item with ID {food_id} does not exist."
         }
         return Response(data, status=status.HTTP_404_NOT_FOUND)
+    if food.user != request.user:
+        raise PermissionDenied(detail={"message":"You do not have the permission to edit this item as you did not create it."})    
     
     if request.method == "GET":
+        user=request.user
+        if user.is_vendor == False:
+            raise PermissionDenied(detail={"message":f"Permission Denied. Only vendors can perform this action"})
         serializer = FoodSerializer(food)
         
         data = {
@@ -94,6 +91,9 @@ def food_detail(request, food_id):
         return Response(data, status=status.HTTP_200_OK)
     
     elif request.method == 'PUT':
+        user=request.user
+        if user.is_vendor == False:
+            raise PermissionDenied(detail={"message":f"Permission Denied. Only vendors can perform this action"})
         serializer = FoodSerializer(food, data=request.data, partial=True)
         if serializer.is_valid():
                 
@@ -111,6 +111,9 @@ def food_detail(request, food_id):
             return Response(data, status=status.HTTP_400_BAD_REQUEST)
         
     elif request.method=="DELETE":
+        user=request.user
+        if user.is_vendor == False:
+            raise PermissionDenied(detail={"message":f"Permission Denied. Only vendors can perform this action"})
         food.delete()
         
         return Response({}, status=status.HTTP_204_NO_CONTENT)
@@ -124,8 +127,11 @@ def food_detail(request, food_id):
 @authentication_classes([BasicAuthentication])
 @permission_classes([IsAuthenticated])
 def menus(request):
+    # get logged in vendors menu and manually add to menu for logged in vendor
     user= request.user
     if request.method == "GET":
+        if user.is_vendor == False:
+            raise PermissionDenied(detail={"message":f"Permission Denied. Only vendors can perform this action"})
         menu = Menu.objects.filter(user=user)
         serializer = MenuSerializer(menu, many=True)
         
@@ -138,6 +144,8 @@ def menus(request):
         return Response(data, status=status.HTTP_200_OK)
     
     elif request.method == 'POST':
+        if user.is_vendor == False:
+            raise PermissionDenied(detail={"message":f"Permission Denied. Only vendors can perform this action"})
         serializer = MenuSerializer(data=request.data)
         if serializer.is_valid(): 
             if "user" in serializer.validated_data.keys():
@@ -150,7 +158,14 @@ def menus(request):
                 raise PermissionDenied(detail={"message":"You do not have the permission to add this item."})
 
             
-
+            check_in_menu=Menu.objects.filter(user=user, food=food)
+            if check_in_menu.exists():
+                data = {
+                    'message' : 'Failed. Food item already in your menu'
+                }
+                return Response(data, status=status.HTTP_400_BAD_REQUEST)
+            else:
+                pass
             menu = Menu.objects.create(user=user, food=food)
             new_serializer = MenuSerializer(menu)
             
@@ -165,33 +180,30 @@ def menus(request):
                 'error'  : serializer.errors
             }
             return Response(data, status=status.HTTP_400_BAD_REQUEST)
-        
-    
-    
-    
-
-    
     
 
 @api_view(['GET', 'DELETE'])
 @authentication_classes([BasicAuthentication])
 @permission_classes([IsAuthenticated])
-def menu_detail(request, item_id):
+def menu_detail(request, menu_id):
 
-    
+    # get, delete an individual menu item
     try:
-        food = Menu.objects.get(id=item_id)
+        food = Menu.objects.get(id=menu_id)
     except Menu.DoesNotExist:
 
         data = {
             'message' : 'failed',
-            'error'  : f"Menu with ID {item_id} does not exist."
+            'error'  : f"Menu with ID {menu_id} does not exist."
         }
         return Response(data, status=status.HTTP_404_NOT_FOUND)
     if food.user != request.user:
         raise PermissionDenied(detail={"message":"You do not have the permission to view this item."})
     
     if request.method == "GET":
+        user=request.user
+        if user.is_vendor == False:
+            raise PermissionDenied(detail={"message":f"Permission Denied. Only vendors can perform this action"})
         serializer = MenuSerializer(food)
         
         data = {
@@ -203,6 +215,9 @@ def menu_detail(request, item_id):
         return Response(data, status=status.HTTP_200_OK)
     
     elif request.method=="DELETE":
+        user=request.user
+        if user.is_vendor == False:
+            raise PermissionDenied(detail={"message":f"Permission Denied. Only vendors can perform this action"})
         food.delete()
         
         return Response({}, status=status.HTTP_204_NO_CONTENT)
